+++ src/main/java/org/mmarini/hilbert/model/Status.java
@@ -45,7 +45,7 @@ import static org.mmarini.hilbert.model.ExtMath.softmax;
  * and the resources' distribution preferences (softmax)
  */
 public class Status {
-    public static final Status ZERO = new Status(0, 0, 0, 0, 0, 0, 0, 0, 0, 0);
+    public static final Status ZERO = new Status(0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0);
 
     /**
      * Returns the status
@@ -53,15 +53,17 @@ public class Status {
      * @param farmers     the number of farmers
      * @param researchers the number of researchers
      * @param educators   the number of educators
+     * @param doctors     the number of doctors
      * @param inactive    the number of inactive
      * @param food        the food resources
      * @param research    the research resources
      * @param education   the education resources
+     * @param health      the health resources
      * @param settlement  the settlement resources
      * @param technology  the technology level
      */
-    public static Status create(int farmers, int researchers, int educators, int inactive,
-                                double food, double research, double education, double settlement,
+    public static Status create(int farmers, int researchers, int educators, int doctors, int inactive,
+                                double food, double research, double education, double health, double settlement,
                                 double technology) {
         if (farmers <= 0) {
             throw new IllegalArgumentException(format("Farmers must be positive (%d)", farmers));
@@ -71,6 +73,9 @@ public class Status {
         }
         if (educators <= 0) {
             throw new IllegalArgumentException(format("Educators must be positive (%d)", educators));
+        }
+        if (doctors <= 0) {
+            throw new IllegalArgumentException(format("Doctors individuals must be positive (%d)", doctors));
         }
         if (inactive <= 0) {
             throw new IllegalArgumentException(format("Inactive individuals must be positive (%d)", inactive));
@@ -87,11 +92,11 @@ public class Status {
         if (settlement <= 0) {
             throw new IllegalArgumentException(format("Settlement must be positive (%e)", settlement));
         }
-        int population = farmers + researchers + educators + inactive;
-        double[] popPrefs = invSoftmax(farmers, researchers, educators, inactive);
-        double[] resPrefs = invSoftmax(food, research, education, settlement);
-        return new Status(population, popPrefs[0], popPrefs[1], popPrefs[2], popPrefs[3],
-                resPrefs[0], resPrefs[1], resPrefs[2], resPrefs[3],
+        int population = farmers + researchers + educators + doctors + inactive;
+        double[] popPrefs = invSoftmax(farmers, researchers, educators, doctors, inactive);
+        double[] resPrefs = invSoftmax(food, research, education, health, settlement);
+        return new Status(population, popPrefs[0], popPrefs[1], popPrefs[2], popPrefs[3], popPrefs[4],
+                resPrefs[0], resPrefs[1], resPrefs[2], resPrefs[3], resPrefs[4],
                 technology);
     }
 
@@ -101,7 +106,7 @@ public class Status {
      * @param population the number of change individuals
      */
     public static Status population(int population) {
-        return new Status(population, 0, 0, 0, 0, 0, 0, 0, 0, 0);
+        return new Status(population, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0);
     }
 
     /**
@@ -114,10 +119,12 @@ public class Status {
         double farmerPrefs = 0;
         double researcherPrefs = 0;
         double educatorPrefs = 0;
+        double doctorPrefs = 0;
         double inactivePrefs = 0;
         double foodPrefs = 0;
         double researchPrefs = 0;
         double educationPrefs = 0;
+        double healthPrefs = 0;
         double settlementPrefs = 0;
         double technology = 0;
         for (Status state : states) {
@@ -125,15 +132,17 @@ public class Status {
             farmerPrefs += state.farmerPrefs;
             researcherPrefs += state.researcherPrefs;
             educatorPrefs += state.educatorPrefs;
+            doctorPrefs += state.doctorPrefs;
             inactivePrefs += state.inactivePrefs;
             foodPrefs += state.foodPrefs;
             researchPrefs += state.researchPrefs;
             educationPrefs += state.educationPrefs;
+            healthPrefs += state.healthPrefs;
             settlementPrefs += state.settlementPrefs;
             technology += state.technology;
         }
-        return new Status(population, farmerPrefs, researcherPrefs, educatorPrefs, inactivePrefs,
-                foodPrefs, researchPrefs, educationPrefs, settlementPrefs, technology);
+        return new Status(population, farmerPrefs, researcherPrefs, educatorPrefs, doctorPrefs, inactivePrefs,
+                foodPrefs, researchPrefs, educationPrefs, healthPrefs, settlementPrefs, technology);
     }
 
     /**
@@ -142,7 +151,7 @@ public class Status {
      * @param value the technology value
      */
     public static Status technology(double value) {
-        return new Status(0, 0, 0, 0, 0, 0, 0, 0, 0, value);
+        return new Status(0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, value);
     }
 
     /**
@@ -156,10 +165,12 @@ public class Status {
     private final double farmerPrefs;
     private final double researcherPrefs;
     private final double educatorPrefs;
+    private final double doctorPrefs;
     private final double inactivePrefs;
     private final double foodPrefs;
     private final double researchPrefs;
     private final double educationPrefs;
+    private final double healthPrefs;
     private final double settlementPrefs;
     private final double technology;
     private final LazyValue<double[]> individuals;
@@ -172,33 +183,37 @@ public class Status {
      * @param farmerPrefs     the farmer preferences
      * @param researcherPrefs the researcher preferences
      * @param educatorPrefs   the educator preferences
+     * @param doctorPrefs     the doctor preferences
      * @param inactivePrefs   the inactive preferences
      * @param foodPrefs       the food preferences
      * @param researchPrefs   the research preferences
      * @param educationPrefs  the education preferences
+     * @param healthPrefs     the health preferences
      * @param settlementPrefs the settlement preferences
      * @param technology      the technology level
      */
-    public Status(int population, double farmerPrefs, double researcherPrefs, double educatorPrefs, double inactivePrefs,
-                  double foodPrefs, double researchPrefs, double educationPrefs, double settlementPrefs, double technology) {
+    public Status(int population, double farmerPrefs, double researcherPrefs, double educatorPrefs, double doctorPrefs, double inactivePrefs,
+                  double foodPrefs, double researchPrefs, double educationPrefs, double healthPrefs, double settlementPrefs, double technology) {
         this.population = population;
         this.farmerPrefs = farmerPrefs;
         this.researcherPrefs = researcherPrefs;
         this.educatorPrefs = educatorPrefs;
+        this.doctorPrefs = doctorPrefs;
         this.inactivePrefs = inactivePrefs;
         this.foodPrefs = foodPrefs;
         this.researchPrefs = researchPrefs;
         this.educationPrefs = educationPrefs;
+        this.healthPrefs = healthPrefs;
         this.settlementPrefs = settlementPrefs;
         this.technology = technology;
         this.individuals = new LazyValue<>(() ->
-                Arrays.stream(softmax(farmerPrefs, researcherPrefs, educatorPrefs, inactivePrefs))
-                        .limit(3)
+                Arrays.stream(softmax(farmerPrefs, researcherPrefs, educatorPrefs, doctorPrefs, inactivePrefs))
+                        .limit(4)
                         .map(a -> a * population)
                         .toArray());
         this.betas = new LazyValue<>(() ->
-                Arrays.stream(softmax(foodPrefs, researchPrefs, educationPrefs, settlementPrefs))
-                        .limit(3)
+                Arrays.stream(softmax(foodPrefs, researchPrefs, educationPrefs, healthPrefs, settlementPrefs))
+                        .limit(4)
                         .toArray()
         );
     }
@@ -208,7 +223,18 @@ public class Status {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
         Status status = (Status) o;
-        return population == status.population && Double.compare(status.farmerPrefs, farmerPrefs) == 0 && Double.compare(status.researcherPrefs, researcherPrefs) == 0 && Double.compare(status.educatorPrefs, educatorPrefs) == 0 && Double.compare(status.inactivePrefs, inactivePrefs) == 0 && Double.compare(status.foodPrefs, foodPrefs) == 0 && Double.compare(status.researchPrefs, researchPrefs) == 0 && Double.compare(status.educationPrefs, educationPrefs) == 0 && Double.compare(status.settlementPrefs, settlementPrefs) == 0 && Double.compare(status.technology, technology) == 0;
+        return population == status.population && Double.compare(status.farmerPrefs, farmerPrefs) == 0 && Double.compare(status.researcherPrefs, researcherPrefs) == 0 && Double.compare(status.educatorPrefs, educatorPrefs) == 0 && Double.compare(status.doctorPrefs, doctorPrefs) == 0 && Double.compare(status.inactivePrefs, inactivePrefs) == 0 && Double.compare(status.foodPrefs, foodPrefs) == 0 && Double.compare(status.researchPrefs, researchPrefs) == 0 && Double.compare(status.educationPrefs, educationPrefs) == 0 && Double.compare(status.healthPrefs, healthPrefs) == 0 && Double.compare(status.settlementPrefs, settlementPrefs) == 0 && Double.compare(status.technology, technology) == 0;
+    }
+
+    public double getDoctorPrefs() {
+        return doctorPrefs;
+    }
+
+    /**
+     * Returns the number of doctors
+     */
+    public double getDoctors() {
+        return individuals.get()[3];
     }
 
     public double getEducationPrefs() {
@@ -262,6 +288,17 @@ public class Status {
         return betas.get()[0];
     }
 
+    public double getHealthPrefs() {
+        return healthPrefs;
+    }
+
+    /**
+     * Returns the health ratio
+     */
+    public double getHealthRatio() {
+        return betas.get()[3];
+    }
+
     public double getInactivePrefs() {
         return inactivePrefs;
     }
@@ -310,7 +347,7 @@ public class Status {
      * Returns the research ratio
      */
     public double getSettlementRatio() {
-        return 1 - betas.get()[0] - betas.get()[1] - betas.get()[2];
+        return 1 - betas.get()[0] - betas.get()[1] - betas.get()[2] - betas.get()[3];
     }
 
     public double getTechnology() {
@@ -324,14 +361,14 @@ public class Status {
      */
     public Status setTechnology(double technology) {
         return technology != this.technology
-                ? new Status(population, farmerPrefs, researcherPrefs, educatorPrefs, inactivePrefs,
-                foodPrefs, researchPrefs, educationPrefs, settlementPrefs, technology)
+                ? new Status(population, farmerPrefs, researcherPrefs, educatorPrefs, doctorPrefs, inactivePrefs,
+                foodPrefs, researchPrefs, educationPrefs, healthPrefs, settlementPrefs, technology)
                 : this;
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(population, farmerPrefs, researcherPrefs, educatorPrefs, inactivePrefs, foodPrefs, researchPrefs, educationPrefs, settlementPrefs, technology);
+        return Objects.hash(population, farmerPrefs, researcherPrefs, educatorPrefs, doctorPrefs, inactivePrefs, foodPrefs, researchPrefs, educationPrefs, healthPrefs, settlementPrefs, technology);
     }
 
     /**
@@ -356,11 +393,11 @@ public class Status {
                 farmerPrefs - popRefsOffset,
                 researcherPrefs - popRefsOffset,
                 educatorPrefs - popRefsOffset,
-                inactivePrefs - popRefsOffset,
+                doctorPrefs, inactivePrefs - popRefsOffset,
                 foodPrefs - resRefsOffset,
                 researchPrefs - resRefsOffset,
                 educationPrefs - resRefsOffset,
-                settlementPrefs - resRefsOffset,
+                healthPrefs, settlementPrefs - resRefsOffset,
                 newTechnology)
                 : this;
     }
@@ -372,10 +409,12 @@ public class Status {
                 .add("farmerPrefs=" + farmerPrefs)
                 .add("researcherPrefs=" + researcherPrefs)
                 .add("educatorPrefs=" + educatorPrefs)
+                .add("doctorPrefs=" + doctorPrefs)
                 .add("inactivePrefs=" + inactivePrefs)
                 .add("foodPrefs=" + foodPrefs)
                 .add("researchPrefs=" + researchPrefs)
                 .add("educationPrefs=" + educationPrefs)
+                .add("healthPrefs=" + healthPrefs)
                 .add("settlementPrefs=" + settlementPrefs)
                 .add("technology=" + technology)
                 .toString();